% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{allhexes}
\alias{allhexes}
\title{Run speed bass on all hexagons and all samples}
\usage{
allhexes(hexes, samples, total, printDets = FALSE)
}
\arguments{
\item{hexes}{Matrix of hexagon land covers. Rows are hexagons, columns are land cover types}

\item{samples}{Matrix of hexagon land covers from random sample. Rows are hexagons, columns are land cover types}

\item{total}{Vector of total land cover. values are individual land cover types}
}
\description{
Run speed bass on all hexagons and all samples
}
