% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASSr-deprecated.R, R/BASS_benefit.R,
%   R/BASS_extract_lcc_cost.R, R/BASS_full_run.R, R/BASS_studyArea_analysis.R
\name{BASSr-deprecated}
\alias{BASSr-deprecated}
\alias{subsample_grts_and_calc_benefit}
\alias{extract_habitat_cost}
\alias{genraster}
\alias{noGRTS_BASS_run}
\alias{getresults_BASS}
\title{BASSr deprecated functions}
\usage{
subsample_grts_and_calc_benefit()

extract_habitat_cost()

genraster()

noGRTS_BASS_run()

getresults_BASS()
}
\description{
\strong{Deprecated} These functions are no longer useful.
}
\seealso{
\link{BASSr-defunct}
}
