% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamstats.R
\name{batch_StreamStatsBC}
\alias{batch_StreamStatsBC}
\title{Batch Basin Characteristics}
\usage{
batch_StreamStatsBC(watersheds, parallel = FALSE, ...)
}
\arguments{
\item{watersheds}{A previously created batch_StreamStatsWS object.}

\item{parallel}{\code{logical} indicating whether to use future_map(). (experimental)}

\item{...}{additional arguments to pass to furrr::future_map().}
}
\value{
A sf data.frame with watersheds and basin characteristics.
}
\description{
Batch Basin Characteristics
}
