% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamstats.R
\name{batch_StreamStatsRBS}
\alias{batch_StreamStatsRBS}
\title{Batch Regression Based Scenarios (RBS)}
\usage{
batch_StreamStatsRBS(data, parallel = FALSE, ...)
}
\arguments{
\item{data}{A previously created batch_StreamStatsBC object.}

\item{parallel}{\code{logical} indicating whether to use future_map().}

\item{...}{additional arguments to pass to furrr::future_map().}
}
\value{
Returns a data.frame with associated regional regression estimates.
}
\description{
Provides the Regression Based Scenarios (RBS).
Uses methods from \insertCite{ries2017streamstats} to generate RBS.
}
\note{
The use of \code{parallel} is likely to not return all of the requests.
}
\examples{
\dontrun{
# Bring in data

#### use batch_StreamStats object

data <- tibble(Lat = c(48.30602, 48.62952, 48.14946),
                 Lon = c(-115.54327, -114.75546, -116.05935),
                   Site = c("Granite Creek", "Louis Creek", "WF Blue Creek"))

data <- data \%>\%
         sf::st_as_sf(coords = c('Lon', 'Lat')) \%>\%
         sf::st_set_crs(4326)


}

}
\references{
{
\insertAllCited{}
}
}
