% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamstats.R
\name{batch_StreamStatsWS}
\alias{batch_StreamStatsWS}
\title{Download Multiple Stream Stats Locations}
\usage{
batch_StreamStatsWS(data, group, crs = 4326, parallel = FALSE, ...)
}
\arguments{
\item{data}{sf data.frame with POINT geometry}

\item{group}{A vector to group by. \code{optional}}

\item{crs}{A \code{numeric} crs value}

\item{parallel}{\code{logical} indicating whether to use future_map(). (experimental)}

\item{...}{additional arguments to pass to furrr::future_map().}
}
\value{
Returns an sf (simple feature) object with associated basin characteristics.
}
\description{
Takes sf point object and returns watershed boundary (sf). Uses state to generate basin
delineation(s).
}
\note{
The use of \code{parallel} is likely to not return all of the requests.
}
\examples{
\dontrun{
# Bring in data

data <- tidyr::tibble(Lat = c(48.30602, 48.62952, 48.14946),
                 Lon = c(-115.54327, -114.75546, -116.05935),
                   Site = c("Granite Creek", "Louis Creek", "WF Blue Creek"))

data <- data \%>\%
         sf::st_as_sf(coords = c('Lon', 'Lat')) \%>\%
         sf::st_set_crs(4326)

three_sites <- batch_StreamStats(data,
                                 group = 'Site',
                                 crs = 4326,
                                 parallel = FALSE)

}
}
