% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOSS_hab_inclpr.R
\name{calculate_BOSS_hab_inlc_pr}
\alias{calculate_BOSS_hab_inlc_pr}
\title{Calculate habitat probabilities}
\usage{
calculate_BOSS_hab_inlc_pr(land_hex, hex_id, stratum_id = NULL, quiet = FALSE)
}
\arguments{
\item{land_hex}{(Spatial) Data frame. Land Cover data by hexagon. If
non-spatial, will be converted to spatial sf data frame using the \code{crs} and
\code{coords} arguments. Requires columns identifying the Hex ID as well as the
Stratum ID (see \code{hex_id} and \code{stratum_id} respectively).}

\item{hex_id}{Column. Identifies hexagon ids (e.g., \code{hex_id} or \code{ET_Index}).}

\item{stratum_id}{Column. Identifies larger area (e.g., \code{StudyAreaID} or
\code{Province}).}

\item{quiet}{Logical. Whether to suppress progress messages.}
}
\description{
Based on Van Wilgenburg et al. 2020
https://doi.org/10.1371/journal.pone.0234494
}
\examples{

calculate_BOSS_hab_inlc_pr(
  land_hex = psu_hexagons,
  hex_id = hex_id)
}
