% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASS_cost_ben.R
\name{calculate_inclusion_probs}
\alias{calculate_inclusion_probs}
\title{BASS cost benefit calculation}
\usage{
calculate_inclusion_probs(
  cost,
  hexagon_benefits,
  HexID,
  StratumID = StudyAreaID,
  benefit_weight = 0.5
)
}
\arguments{
\item{cost}{data frame with costs for each hexagon in a RawCost format}

\item{hexagon_benefits}{data frame with benefits for each hexagon}

\item{HexID}{column of hexagon ids}

\item{StratumID}{Column of larger area id}

\item{benefit_weight}{The weight assigned to benefit in the selection probabilities.0.5 is equal weighting of cost and benefits. 1.0 is zero weighting to cost.}
}
\value{
A data frame with full inclusion probabilities for each raster.
}
\description{
Caculate the cost benefits and inclusion probabilities
}
