% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternative_inclpr.R
\name{calculate_PPS_hab_inlc_pr}
\alias{calculate_PPS_hab_inlc_pr}
\title{Calculate Propbability Proportional to Size (PPS) inclusion probabilities}
\usage{
calculate_PPS_hab_inlc_pr(
  land_hex,
  hex_id = hex_id,
  stratum_id = NULL,
  quiet = FALSE
)
}
\arguments{
\item{land_hex}{(Spatial) Data frame. Land Cover data by hexagon. If
non-spatial, will be converted to spatial sf data frame using the \code{crs} and
\code{coords} arguments. Requires columns identifying the Hex ID as well as the
Stratum ID (see \code{hex_id} and \code{stratum_id} respectively).}

\item{hex_id}{Column. Identifies hexagon IDs (e.g., default \code{hex_id}).}

\item{stratum_id}{Column. Identifies larger area (e.g., \code{StudyAreaID} or
\code{Province}).}

\item{quiet}{Logical. Whether to suppress progress messages.}
}
\value{
tibble with selection weights from PPS
}
\description{
See for more information:
https://www150.statcan.gc.ca/n1/en/pub/12-001-x/2011001/article/11450-eng.pdf?st=0oyBln55
or
https://en.wikipedia.org/wiki/Probability-proportional-to-size_sampling
}
\examples{
 calculate_PPS_hab_inlc_pr(land_hex = psu_hexagons)
}
