% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternative_inclpr.R
\name{calculate_z_scores}
\alias{calculate_z_scores}
\title{Calculate z-scores for each hexagon by sum of individual z scores}
\usage{
calculate_z_scores(land_hex, hex_id, stratum_id = NULL, quiet = FALSE)
}
\arguments{
\item{land_hex}{(Spatial) Data frame. Land Cover data by hexagon. If
non-spatial, will be converted to spatial sf data frame using the \code{crs} and
\code{coords} arguments. Requires columns identifying the Hex ID as well as the
Stratum ID (see \code{hex_id} and \code{stratum_id} respectively).}

\item{hex_id}{Column. Identifies hexagon IDs (e.g., \code{hex_id} or \code{ET_Index}).}

\item{stratum_id}{Column. Identifies larger area (e.g., \code{StudyAreaID} or
\code{Province}).}

\item{quiet}{Logical. Whether to suppress progress messages.}
}
\value{
data frame
}
\description{
Calculate z-scores for each hexagon by sum of individual z scores
}
