% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean4Bass.R
\name{clean_forBass}
\alias{clean_forBass}
\title{Clean habitat data for BASS}
\usage{
clean_forBass(
  df,
  s,
  id_col = StudyAreaID,
  f_vec = brandtStudyAreas_list,
  appended = ""
)
}
\arguments{
\item{df}{Habitat data frame}

\item{s}{String to search and replace with 'LC'}

\item{id_col}{Column name to use for filtering from f_vec}

\item{f_vec}{Vector of id to include}

\item{appended}{String to append to end of LandCover Code}
}
\value{

}
\description{
This is a general function to clean land cover data for BASS functions
}
