% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASSr-package.R
\docType{data}
\name{clrfile}
\alias{clrfile}
\title{rgb colour codes to plot 2015 National Land Cover.}
\format{
A data frame with 20 rows and 6 variables:
\describe{
  \item{Value}{price, in US dollars}
  \item{red}{weight of the diamond, in carats}
  \item{green}{weight of the diamond, in carats}
  \item{blue}{weight of the diamond, in carats}
  \item{rgb}{weight of the diamond, in carats}
  \item{LCC_NAME}{weight of the diamond, in carats}

}
}
\source{
\url{https://open.canada.ca/data/en/dataset/4e615eae-b90c-420b-adee-2ca35896caf6}
}
\usage{
clrfile
}
\description{
A dataset containing red, blue, green values associated with
all the land cover types found in the 2015 Canadian National
Land Cover Classification
}
\keyword{datasets}
