% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_common_docs.R
\name{common_docs}
\alias{common_docs}
\title{Common arguments and documentation for various functions}
\arguments{
\item{att_sf}{Spatial Data frame. sf attributes table.}

\item{stratum_id}{Column. Identifies larger area (e.g., \code{StudyAreaID} or
\code{Province})}

\item{hex_id}{Columns. Identifies hexagon ids (e.g., \code{hex_id} or \code{ET_Index})}

\item{costs}{Data frame. Costs for each hexagon in a RawCost format.}
}
\description{
Common arguments and documentation for various functions
}
\details{
Use \verb{@inheritParams common_docs} to include the above in any function
documentation with a matching argument (will only include matching args)
}
\keyword{internal}
