% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASS_cost_1.R
\docType{data}
\name{cost_vars}
\alias{cost_vars}
\title{Variables for cost estimation}
\format{
A list with 18 variables used to calculate the cost model for N Ontario:
\describe{
  \item{truck_cost_per_day}{Cost of truck use per day per crew, in dollars}
  \item{truck_n_crews}{Number of crews for truck surveys}
  \item{truck_arus_per_crew_per_day}{number of arus deployed per crew per day}
  \item{atv_cost_per_day}{Cost of ATV use per day per crew, in dollars}
  \item{atv_n_crews}{Number of crews for ATV surveys}
  \item{atv_arus_per_crew_per_day}{Number of arus deployed per crew per day}
  \item{helicopter_cost_per_hour}{Cost of helicopter rental per hour, in dollars}
  \item{helicopter_max_km_from_base}{Maximum range of helicopter from fuelling base, in kilometres}
  \item{helicopter_base_setup_cost_per_km}{Cost of setting up base for helicopter use with distance from airport, in km}
  \item{helicopter_l_per_hour}{Helicopter fuel usage per hour, in litres}
  \item{helicopter_crew_size}{Helicopter crew size}
  \item{helicopter_aru_per_person_per_day}{Number of arus deployed per day per person}
  \item{helicopter_relocation_speed}{Speed of movement of helicopter when relocating, in km per hour}
  \item{helicopter_airport_cost_per_l}{Cost of heli fuel from an airport, in dollars per litre}
  \item{helicopter_base_cost_per_l}{Cost of heli fuel from a basecamp, in dollars per litre}
  \item{helicopter_2nd_base_cost_per_l}{Cost of heli fuel from a a remote fuel cache, in dollars per litre}
  \item{helicopter_hours_flying_within_sa_per_day}{Number of hours a helicopter spends flying in the study area per day.}
  ...
}
}
\source{
Advice from Rich Russell
}
\usage{
cost_vars
}
\description{
Cost variables for estimate_cost_study_area
}
\keyword{datasets}
