% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalSSU_gen.R
\name{create_sites}
\alias{create_sites}
\title{Create sampling sites within hexagons}
\usage{
create_sites(hex, spacing = NULL, n = NULL, hex_id = hex_id)
}
\arguments{
\item{hex}{Spatial Data frame. Hexagon grid across sampling region.
Requires column identifying the hexagon IDs (see \code{hex_id})}

\item{spacing}{Numeric. Distance between sites. Units
are assumed to be those of \code{hex} spatial data frame.}

\item{n}{Numeric. \emph{Approximate} number of sites to create within a hex grid.}

\item{hex_id}{Column. Identifies hexagon IDs (e.g., \code{hex_id} or \code{ET_Index}).}
}
\value{
Spatial data frame of sites as points.
}
\description{
Creates a grid of sites within the hexagon grid cells created by
\code{create_hexes()}. These sites can then be sampled with \code{sample_sites()}.
}
\examples{
# Get sites by exact within-hex distances
sites_sp <- create_sites(psu_hexagons, spacing = 5)

# Get sites by approximate number of points (but equal spacing among hexes)
sites_n <- create_sites(psu_hexagons, n = 61)

# Same number of sites, but in slightly different spots, because creating by
# n maximizes spacing, but creating by spacing using the exact spacing
# specified.

library(ggplot2)
ggplot() +
  geom_sf(data = psu_hexagons) +
  geom_sf(data = sites_sp, size = 0.5, colour = "red") +
  geom_sf(data = sites_n, size = 0.5, colour = "blue")
}
