% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downweight.R
\name{downweight_selection_pr}
\alias{downweight_selection_pr}
\title{Adjust selection weighting}
\usage{
downweight_selection_pr(
  sample_locs,
  scalingFactor,
  sigma_value,
  selection_column = NULL,
  fun = "cauchy",
  existing_sampling = NULL,
  dmat = NULL
)
}
\arguments{
\item{sample_locs}{sf object with the selection weighting column (polygons or points)}

\item{scalingFactor}{scaling factor to downweight}

\item{sigma_value}{sigma value of distribution effect for existing sampling. Larger value means sampling has wider effect}

\item{selection_column}{Column with sampling weights to be adjusted. If null return only the weights.}

\item{fun}{Type of decay function. Current options are 'cauchy', 'normal' or 'exp'}

\item{existing_sampling}{exisint sampling to down weight around (points)}

\item{dmat}{distance matrix from sample locations (rows) to existing sampling (columns)}
}
\value{
data.frame sample_locs with downweights and adjusted selection probabilities included.
If selection column is null return a vector of the downweights alone.
}
\description{
Adjust selection weighting
}
\examples{
# downweight_selection_pr(BASSr::all_study_areas[1:10,], BASSr::all_study_areas[20:30,], scalingFactor = 0.1, sigma_value = 1e5, fun = 'cauchy', selection_column = NULL)
}
