% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASS_cost_1.R
\name{estimate_cost_study_area}
\alias{estimate_cost_study_area}
\title{Cost model estimate}
\usage{
estimate_cost_study_area(
  narus,
  StudyAreas,
  pr,
  sr,
  dist_base_sa,
  dist_airport_sa,
  dist2airport_base,
  AirportType,
  vars
)
}
\arguments{
\item{narus}{The number of ARUs to deploy in the study area}

\item{StudyAreas}{Tibble with study area information and distances}

\item{pr}{Column with primary road buffer proportion of study area}

\item{sr}{Column with secondary Road proportion of study area ( should not include primary road area)}

\item{dist_base_sa}{Column with distance between basecamp and study area}

\item{dist_airport_sa}{Column with distance between airport and study area}

\item{dist2airport_base}{Column with distance between airport and base camp}

\item{AirportType}{Column with nearest airport type}

\item{vars}{}
}
\value{

}
\description{
Cost model estimate
}
