% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downweight.R
\name{gen_dist_mat}
\alias{gen_dist_mat}
\title{Wrapper for st_distance from sf that uses centroids of sample locations}
\usage{
gen_dist_mat(sample_locs, existing_sampling)
}
\arguments{
\item{sample_locs}{sf object with the selection weighting column (polygons or points)}

\item{existing_sampling}{exisint sampling to down weight around (points)}
}
\value{
sf_distance matrix of distances with sample locations as rows and existing sampling as columns
}
\description{
Wrapper for st_distance from sf that uses centroids of sample locations
}
\examples{
gen_dist_mat(BASSr::all_study_areas[1:10,], BASSr::all_study_areas[20:30,])
}
