% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_Points.R
\name{generate_Points_around_centroid}
\alias{generate_Points_around_centroid}
\title{Extract Point Count or multiple locations within a sample unit.}
\usage{
generate_Points_around_centroid(df, crs_ = NULL)
}
\arguments{
\item{df}{data.table or tibble or sf object that contains sample units to add points to.}

\item{crs_}{CRS to add to the points. Should be based on the centroids. Does not to be specificied if df is an 'sf' object.}
}
\value{
Returns an sf object with points within provided hexagons
}
\description{
Extract Point Count or multiple locations within a sample unit.
}
