% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASS_extract_lcc_cost.R
\name{genraster}
\alias{genraster}
\title{Generatate Raster}
\usage{
genraster(
  id_col,
  studyareas,
  id,
  hab_rast_location,
  writeout = T,
  outpath = "output"
)
}
\arguments{
\item{id_col}{Column to use}

\item{studyareas}{Study area list}

\item{id}{Id of study area to extract}

\item{hab_rast_location}{Location of habitat file}

\item{writeout}{Logical - write output}

\item{outpath}{Location of output}
}
\description{
Generatate Raster
}
