% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalSSU_gen.R
\name{genSSU}
\alias{genSSU}
\title{genSSU. Generate ssu layer aligned with hexagon PSU}
\usage{
genSSU(h, spacing)
}
\arguments{
\item{h}{Hexagon - sf polygon}

\item{spacing}{distance between points}
}
\value{
SSU points as an sf object
}
\description{
genSSU. Generate ssu layer aligned with hexagon PSU
}
