% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamstats.R
\name{get_rbs}
\alias{get_rbs}
\title{regression based scenarios}
\usage{
get_rbs(state, wkID, group)
}
\arguments{
\item{state}{A character abb of a state}

\item{wkID}{A previously created wkID}

\item{group}{A group in the data.frame}
}
\description{
regression based scenarios
}
