% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpoints.R
\name{get_StreamStats_points}
\alias{get_StreamStats_points}
\title{Get Stream Stats Stream Tile}
\usage{
get_StreamStats_points(state)
}
\arguments{
\item{state}{Must be a single state abbreviation or full name}
}
\value{
A mapedit result
}
\description{
Get Stream Stats Stream Tile
}
