% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASS_studyArea_analysis.R
\name{getresults_BASS-deprecated}
\alias{getresults_BASS-deprecated}
\title{Get the results from a BASS grts run}
\arguments{
\item{grts_output}{Hypothetical sample set}

\item{study_area_results}{Study area results of BASS}

\item{nARUs}{Number of ARUs to deploy}
}
\description{
Get the results from a BASS grts run
}
