% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASS_studyArea_analysis.R
\name{getresults_BASS}
\alias{getresults_BASS}
\title{Get the results from a BASS grts run}
\usage{
getresults_BASS(grts_output, study_area_results, nARUs)
}
\arguments{
\item{grts_output}{Hypothetical sample set}

\item{study_area_results}{Study area results of BASS}

\item{nARUs}{Number of ARUs to deploy}
}
\value{

}
\description{
Get the results from a BASS grts run
}
