% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASS_cost_1.R
\name{getroaddensity}
\alias{getroaddensity}
\title{Calculate road density in a hexagon}
\usage{
getroaddensity(hexes, sa, pr, sr, wr, r, idcol, ...)
}
\arguments{
\item{hexes}{Hexagon layer}

\item{sa}{hexagon id to extract road info from}

\item{pr}{Primary roads (truck roads) buffer. Should be a polygon layer}

\item{sr}{Secondary (atv roads) buffer. Should be a polygon layer}

\item{wr}{Winter roads buffer. Should be a polygon layer}

\item{r}{Total roads buffer. Should be polygon}

\item{idcol}{Column with hexagon ids}

\item{...}{You can include multisession with the furrr package. Needs to include Multicor=T}
}
\value{
Returns a tibble with area covered by each road types and their proportion of the study area
}
\description{
This function uses pregenerated buffers to calculate the proportion of a hexagon covered by 3 road types and total.
}
