% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASSr-package.R
\docType{data}
\name{lcc2015_codes}
\alias{lcc2015_codes}
\title{2015 National Landcover Classification Table}
\format{
A data frame with 19 rows and 2 variables:
\describe{
  \item{LCC_CODE}{Land cover code}
  \item{LCC_NAME}{Land cover name}
  ...
}
}
\source{
\url{https://open.canada.ca/data/en/dataset/4e615eae-b90c-420b-adee-2ca35896caf6}
}
\usage{
lcc2015_codes
}
\description{
Key for land cover codes and names
}
\keyword{datasets}
