% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASS_full_run.R
\name{noGRTS_BASS_run-deprecated}
\alias{noGRTS_BASS_run-deprecated}
\title{A calculate BASS from random samples}
\arguments{
\item{samples}{Hypothetical sample set}

\item{num_runs}{The number of times to draw random samples from hexagons}

\item{n_samples}{The number of samples to draw in each sample}

\item{costs}{the cost table for each hexagon id}
}
\description{
A calculate BASS from random samples
}
