% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASS_full_run.R
\name{noGRTS_BASS_run}
\alias{noGRTS_BASS_run}
\title{A calculate BASS from random samples}
\usage{
noGRTS_BASS_run(
  grts_output,
  num_runs,
  nsamples,
  att,
  att.sp,
  cost,
  return_all = FALSE,
  seed_ = as.integer(Sys.time())
)
}
\arguments{
\item{grts_output}{Hypothetical sample set}

\item{num_runs}{The number of times to draw random samples from hexagons}

\item{nsamples}{The number of samples to draw in each sample}

\item{att}{the attribute table correctly formatted}

\item{att.sp}{the shape file from the attribute table}

\item{cost}{the cost table for each hexagon id}

\item{return_all}{return each piece of BASS implementation}

\item{seed_}{random seed to use for random grts samples}
}
\value{
a table with inclusion probabilities
}
\description{
A calculate BASS from random samples
}
