% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{oppositeSigns}
\alias{oppositeSigns}
\title{Opposite signs True or false}
\usage{
oppositeSigns(x, y)
}
\arguments{
\item{x}{a double value}

\item{y}{an integer}
}
\description{
Opposite signs True or false
}
