% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASS_cost_1.R
\name{prepare_cost}
\alias{prepare_cost}
\title{Propare hexagons for cost calculations}
\usage{
prepare_cost(
  truck_roads,
  atv_roads,
  winter_roads,
  all_roads,
  airports,
  basecamps,
  hexagons,
  idcol_,
  calc_roads = T,
  airport_cols = c("NAME", "AIRPORT_TY", "OGF_ID"),
  basecamp_cols = c("OFFICIAL_N", "OGF_ID", "CLASS_SUBT")
)
}
\arguments{
\item{truck_roads}{Primary roads (truck roads) buffer. Should be a polygon layer. Only used if calculating road estimates.}

\item{atv_roads}{Secondary (atv roads) buffer. Should be a polygon layer. Only used if calculating road estimates.}

\item{winter_roads}{Winter roads buffer. Should be a polygon layer. Not currently in use.}

\item{all_roads}{Total roads buffer. Should be polygon.  Not currently in use.}

\item{airports}{Airport locations. Should be a polygon layer.}

\item{basecamps}{Basecamp locations. Should be a polygon layer.}

\item{hexagons}{Hexagon layer}

\item{idcol_}{Column with hexagon ids}

\item{calc_roads}{Logical. Should you calculate roads or are they already included in hexagon layer}

\item{airport_cols}{Columns to use extract airport info. See examples. Should be length of 3.}

\item{basecamp_cols}{Columns to use extract basecamp info. See examples. Should be length of 3.}
}
\value{

}
\description{
Propare hexagons for cost calculations
}
\examples{
prepare_cost(truck_roads = NA, atv_roads = NA, winter_roads = NA, all_roads = NA, airports = airports_official, basecamps = tourism, hexagons =  study_area_hexagons_in_brandt \%>\%
left_join(road_info, by = c("StudyAreaID" = "StudyArea")),idcol_ = StudyAreaID, calc_roads = F, airport_cols =  c("NAME", "AIRPORT_TY", "OGF_ID"),
  basecamp_cols = c("OFFICIAL_N", "OGF_ID", "CLASS_SUBT") )
}
