% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASS_benefit.R
\name{prepare_hab_long}
\alias{prepare_hab_long}
\title{Summarize land cover in attribute and add it to a longer table}
\usage{
prepare_hab_long(att)
}
\arguments{
\item{att}{attribute table}
}
\value{

}
\description{
Summarize land cover in attribute and add it to a longer table
}
