% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{psu_costs}
\alias{psu_costs}
\title{Dummy costs data}
\format{
A data frame with 33 rows and 27 columns
\describe{
\item{hex_id}{ID of the hex}
\item{province}{Province code for that hex}
\item{water}{Whether that hex is in water or not}
\item{area}{Area of the hex in m2}
\item{pr-total_heli_cost}{Specific costs for each hex (see ?\code{estimate_cost_study_area})}
\item{narus}{Number of ARUs to be deployed}
\item{RawCost}{Total raw cost of sampling this hex}
}
}
\source{
Data generated in data-raw/data_create_study_area.R
}
\usage{
psu_costs
}
\description{
Dummy costs data
}
\keyword{datasets}
