% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{psu_hex_dirty}
\alias{psu_hex_dirty}
\title{Dummy hex data to be cleaned}
\format{
A spatial data frame with 33 features and 9 fields
\describe{
\item{hex_id}{ID of the hex}
\item{province}{Province code for that hex}
\item{water}{Whether that hex is in water or not}
\item{CLC...}{Land cover columns}
\item{x}{Geometry}
}
}
\source{
Data generated in data-raw/data_create_study_area.R
}
\usage{
psu_hex_dirty
}
\description{
Dummy hex data to be cleaned
}
\keyword{datasets}
