% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Full_Run_w_SampleSelect.R
\name{run_full_BASS_w_selection}
\alias{run_full_BASS_w_selection}
\title{Full BASSr run with sample selection}
\usage{
run_full_BASS_w_selection(
  study_areas_hab_cost,
  sample_units_w_hab,
  Number_of_Study_areas,
  Number_of_sample_units,
  Size_of_HSS,
  Number_of_HSS,
  Weight_of_benfit,
  LandCoverType,
  RemovedLayers_,
  Area_of_interest,
  RandomSeed,
  calculate_benefits = list(SA = F, SU = F),
  only_calculate_benefits = list(SA = F, SU = F),
  benefit_dfs = list(SA = NULL, SU = NULL)
)
}
\arguments{
\item{study_areas_hab_cost}{A list of study area hexagons, costs, and habitat characteristics. See naming above.}

\item{Number_of_Study_areas}{Number of Study Areas to select}

\item{Number_of_sample_units}{Number of Sample Units to select}

\item{Size_of_HSS}{Size of they Hypothetical Sample Set}

\item{Number_of_HSS}{Number of iterations in the hypothetical sample set}

\item{Weight_of_benfit}{Weight of benefit in selection probabilties.}

\item{LandCoverType}{String with identifer for both the land cover hexagons and their code within that tibble}

\item{RemovedLayers_}{Layers to remove from benefit calculation must be in format of c(-var1, -var2, -var3)}

\item{RandomSeed}{}
}
\value{
List of sample units and a summary comparing land cover to the local area
}
\description{
Running this function will run BASS on both study areas and sample units within those study areas.
It will return a selection of sample units with associated costs and habitats.
}
