% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSU_selection_methods.R
\name{select_ssu}
\alias{select_ssu}
\title{Select SSU}
\usage{
select_ssu(
  SSUs,
  PSU_id,
  SSU_id,
  type,
  nsamples,
  os,
  cluster_size,
  ARUonly,
  seed_,
  min_dist,
  useGRTS
)
}
\arguments{
\item{SSUs}{sf dataframe of points showing SSU locations}

\item{PSU_id}{Lazy named column showing PSU id}

\item{SSU_id}{String showing id column for SSU id}

\item{type}{String. Method to select SSUs. Currently one of
\itemize{
\item "Cluster"
\item "GTRS" - Depreciated. Use Random with useGRTS = TRUE
\item "ShortestPath"
\item "Random"
}}

\item{nsamples}{Integer total number of points to select}

\item{os}{Double. Proportion of oversample}

\item{cluster_size}{Integer. Number of points per cluster. Not used for GRTS or Random}

\item{ARUonly}{Logical. Used to flag if only ARUs are going to be used.}

\item{seed_}{Random seed}

\item{min_dist}{Numeric. Minimum distance between points}

\item{useGRTS}{Logical. Should the program be run using GRTS? Works for Cluster or Random}
}
\value{
Returns a data frame of points selected from the SSUs object. Except for Shortest path, which
returns a list of the points on the path and the original points selected to create the path.
}
\description{
Selection methods for processing SSU selection using GRTS, random sampling,
clustering, or shortest path methods.
}
