% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{speedbass}
\alias{speedbass}
\title{The internal BASSr benefit algorithm}
\usage{
speedbass(hex, w, sample, total, printDets = FALSE)
}
\arguments{
\item{hex}{A vector of land cover values}

\item{w}{A vector of weights for each land cover value}

\item{sample}{a vector of land cover values from random sample}

\item{total}{a vector of land cover values from total of study area}

\item{printDets}{logical - should you print the details - messy for now.}
}
\description{
The internal BASSr benefit algorithm
}
