% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ssu_land_cover}
\alias{ssu_land_cover}
\title{Dummy SSU land cover}
\format{
A data frame with 3003 rows and 10 columns
\describe{
\item{hex_id}{ID of the hex}
\item{ssuID}{ID of subsampling unit (hex)}
\item{HexArea}{Area of the SSU hex}
\item{LC...}{Land cover columns}
\item{province}{Province code for that hex}
}
}
\source{
Data generated in data-raw/data_create_study_area.R
}
\usage{
ssu_land_cover
}
\description{
Dummy SSU land cover
}
\keyword{datasets}
