% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ssu_points}
\alias{ssu_points}
\title{Dummy SSU points}
\format{
A spatial data frame with 3003 features and 3 fields
\describe{
\item{geometry}{Geometry}
\item{hex_id}{ID of the hex}
\item{ssuID}{ID of subsampling unit (hex)}
\item{province}{Province code for that hex}
}
}
\source{
Data generated in data-raw/data_create_study_area.R
}
\usage{
ssu_points
}
\description{
Dummy SSU points
}
\keyword{datasets}
