% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASS_benefit.R
\name{subsample_grts_and_calc_benefit-deprecated}
\alias{subsample_grts_and_calc_benefit-deprecated}
\title{Subsample GRTS and calculate benefit}
\arguments{
\item{n_samples}{Number of Samples}

\item{num_runs}{Number of iterations}

\item{grts_file}{grts file to run}

\item{land_hex}{Att frame}

\item{quick}{run it using CPP quick calc}
}
\description{
Subsample GRTS and calculate benefit
}
