% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASS_benefit.R
\name{subsample_grts_and_calc_benefit}
\alias{subsample_grts_and_calc_benefit}
\title{Subsample GRTS and calculate benefit}
\usage{
subsample_grts_and_calc_benefit(nsamples, num_runs, grts_file, att, q = T)
}
\arguments{
\item{nsamples}{Number of Samples}

\item{num_runs}{Number of iterations}

\item{grts_file}{grts file to run}

\item{att}{Att frame}

\item{q}{run it using CPP quick calc}
}
\value{

}
\description{
Subsample GRTS and calculate benefit
}
