% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASSr-defunct.R, R/Full_Run_w_SampleSelect.R
\name{BASSr-defunct}
\alias{BASSr-defunct}
\alias{clean_forBass}
\alias{run_full_BASS_w_selection}
\title{BASSr defunct functions}
\usage{
clean_forBass(...)

run_full_BASS_w_selection()
}
\arguments{
\item{...}{Original function arguments}
}
\description{
BASSr defunct functions
}
