% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASSr_defunct.R
\name{BASSr-defunct}
\alias{BASSr-defunct}
\alias{clean_forBass}
\title{BASSr defunct functions}
\usage{
clean_forBass(...)
}
\arguments{
\item{...}{Original function arguments}
}
\description{
BASSr defunct functions
}
