% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASSr-deprecated.R, R/BASS_extract_lcc_cost.R
\name{BASSr-deprecated}
\alias{BASSr-deprecated}
\alias{extract_habitat_cost}
\alias{genraster}
\title{BASSr deprecated functions}
\usage{
extract_habitat_cost(
  number_iterations,
  n_samples_per_iter,
  sample_hexes,
  study_area_hexes,
  id,
  id_col,
  hab_rast_location,
  shape_file_list,
  return_all_,
  nARUs = 30,
  hexid_col = SampleUnitID,
  calc_cost = F,
  calc_hab = F,
  write_hexes = F,
  load_hexes = T,
  rds.loc = "output",
  sa.rast.loc = "output",
  quick = T
)

genraster(
  id_col,
  studyareas,
  id,
  hab_rast_location,
  writeout = T,
  outpath = "output"
)
}
\description{
\strong{Deprecated} These functions are no longer useful.
}
\seealso{
\link{BASSr-defunct}
}
