% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASSr-deprecated.R, R/BASS_benefit.R,
%   R/BASS_extract_lcc_cost.R, R/BASS_full_run.R, R/BASS_studyArea_analysis.R
\name{BASSr-deprecated}
\alias{BASSr-deprecated}
\alias{subsample_grts_and_calc_benefit}
\alias{extract_habitat_cost}
\alias{genraster}
\alias{noGRTS_BASS_run}
\alias{getresults_BASS}
\title{BASSr deprecated functions}
\usage{
subsample_grts_and_calc_benefit(
  n_samples,
  num_runs,
  grts_file,
  land_hex,
  quick = T
)

extract_habitat_cost(
  number_iterations,
  n_samples_per_iter,
  sample_hexes,
  study_area_hexes,
  id,
  id_col,
  hab_rast_location,
  shape_file_list,
  return_all_,
  nARUs = 30,
  hexid_col = SampleUnitID,
  calc_cost = F,
  calc_hab = F,
  write_hexes = F,
  load_hexes = T,
  rds.loc = "output",
  sa.rast.loc = "output",
  quick = T
)

genraster(
  id_col,
  studyareas,
  id,
  hab_rast_location,
  writeout = T,
  outpath = "output"
)

noGRTS_BASS_run(
  land_hex,
  samples,
  num_runs,
  n_samples,
  costs,
  crs = 4326,
  coords = c("lon", "lat"),
  quiet = FALSE
)

getresults_BASS(grts_output, study_area_results, nARUs)
}
\description{
\strong{Deprecated} These functions are no longer useful.
}
\seealso{
\link{BASSr-defunct}
}
