% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASS_benefit.R
\name{calculate_benefit}
\alias{calculate_benefit}
\title{Title}
\usage{
calculate_benefit(
  grts_res,
  HexID,
  att_long,
  output = "all",
  quick = F,
  non_random_set = NULL
)
}
\arguments{
\item{grts_res}{Results from GRTS random sample set.}

\item{HexID}{column for hexagons}

\item{att_long}{Data frame with habitat information, in long format}

\item{output}{if not using quick. Output format}

\item{quick}{should you run it fast?}

\item{non_random_set}{Set of hexagons to include as a non randomly selected set}
}
\value{

}
\description{
Title
}
