% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASS_benefit.R
\name{calculate_benefit}
\alias{calculate_benefit}
\title{Calculate the benefit of a hexagon from  grts results.}
\usage{
calculate_benefit(
  land_hex,
  samples,
  hex_id,
  stratum_id = NULL,
  non_random_set = NULL,
  land_cover_weights = NULL,
  crs = 4326,
  coords = c("lon", "lat"),
  quiet = FALSE
)
}
\arguments{
\item{samples}{(Spatial) Data frame. Results from \code{draw_random_samples()}.}

\item{non_random_set}{Set of hexagons to include as a non randomly selected
set}

\item{land_cover_weights}{Data table with 'lc' and 'weights' column. Or could
be NULL to have all equal.}
}
\description{
Calculate the benefit of a hexagon from  grts results.
}
\examples{

# Using example data psu_hexagons and psu_samples

calculate_benefit(
  land_hex = psu_hexagons,
  samples = psu_samples,
  hex_id = hex_id,
  non_random_set = c("SA_0009", "SA_0022", "SA_0047", "SA_0052"))

# Specify a non-random set

calculate_benefit(
 land_hex = psu_hexagons,
 samples = psu_samples,
 hex_id = hex_id,
 non_random_set = c("SA_0009", "SA_0022", "SA_0047", "SA_0052"))

# Without GRTS

non_grts_samples <- draw_random_samples(
  land_hex = psu_hexagons,
  num_runs = 3,
  n_samples = 10,
  use_grts = FALSE)

calculate_benefit(
 land_hex = psu_hexagons,
 samples = non_grts_samples,
 hex_id = hex_id)



}
