% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOSS_hab_inclpr.R
\name{calculate_BOSS_hab_inlc_pr}
\alias{calculate_BOSS_hab_inlc_pr}
\title{Calculate habitat probabilities}
\usage{
calculate_BOSS_hab_inlc_pr(land_hex, hex_id, stratum_id = NULL, quiet = FALSE)
}
\arguments{
\item{stratum_id}{}
}
\description{
Based on Van Wilgenburg et al. 2020
https://doi.org/10.1371/journal.pone.0234494
}
\examples{

calculate_BOSS_hab_inlc_pr(
  land_hex = psu_hexagons,
  hex_id = hex_id)
}
