% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASS_cost_ben.R
\name{calculate_inclusion_probs}
\alias{calculate_inclusion_probs}
\title{BASS cost benefit calculation}
\usage{
calculate_inclusion_probs(
  benefits,
  costs,
  hex_id = hex_id,
  stratum_id = NULL,
  omit_flag = NULL,
  benefit_weight = 0.5
)
}
\arguments{
\item{benefits}{Spatial Data frame. Benefits associated with each hexagon
(output of \code{calculate_benefits()}).}

\item{costs}{Data frame. Costs for each hexagon in a RawCost format.}

\item{hex_id}{Column. Identifies hexagon IDs (e.g., default \code{hex_id}).}

\item{stratum_id}{Column. Identifies larger area (e.g., \code{StudyAreaID} or
\code{Province}).}

\item{omit_flag}{Column identifying hexes to omit (e.g., water hexes).}

\item{benefit_weight}{Numeric. Weight assigned to benefit in the selection
probabilities. 0.5 is equal weighting of cost and benefits. 1.0 is zero
weighting to cost. Default 0.5.}
}
\value{
A data frame with full inclusion probabilities for each hex.
}
\description{
Calculate the cost-benefits and inclusion probabilities.
}
\examples{

b <- calculate_benefit(land_hex = psu_hexagons,
                       samples = psu_samples)

inc <- calculate_inclusion_probs(
  benefits = b,
  costs = psu_costs)

# Omit water hexes (identified by column `water`)

inc <- calculate_inclusion_probs(
  benefits = b,
  costs = psu_costs,
  omit_flag = water)
}
