% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASS_cost_ben.R
\name{calculate_inclusion_probs}
\alias{calculate_inclusion_probs}
\title{BASS cost benefit calculation}
\usage{
calculate_inclusion_probs(cost, hexagon_benefits, HexID)
}
\arguments{
\item{cost}{data frame with costs for each hexagon in a RawCost format}

\item{hexagon_benefits}{data frame with benefits for each hexagon}

\item{HexID}{column of hexagon ids}
}
\value{
A data frame with full inclusion probabilities for each raster.
}
\description{
Caculate the cost benefits and inclusion probabilities
}
