% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOSS_hab_inclpr.R
\name{calculate_z_scores}
\alias{calculate_z_scores}
\title{Calculate z-scores for each hexagon by sum of individual z scores}
\usage{
calculate_z_scores(att_sf, hex_id, stratum_id = NULL, quiet = FALSE)
}
\arguments{
\item{quiet}{}
}
\description{
Calculate z-scores for each hexagon by sum of individual z scores
}
