% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean4Bass.R
\name{clean_land_cover}
\alias{clean_land_cover}
\title{Clean land cover habitat data}
\usage{
clean_land_cover(att_sf, pattern = "CLC15_", append = "", quiet = FALSE)
}
\arguments{
\item{att_sf}{Spatial Data frame. sf attributes table.}

\item{pattern}{Character. Pattern to match and replace with 'LC'}

\item{append}{Character. Text to append to end of land cover code}
}
\description{
This is a general function to clean land cover columns.
}
\examples{

psu_hex_clean <- clean_land_cover(psu_hex_dirty, pattern = "CLC0013_")


}
