% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_common_docs.R
\name{common_docs}
\alias{common_docs}
\title{Common arguments and documentation for various functions}
\arguments{
\item{num_runs}{Numeric. Number of times to draw random samples.}

\item{land_hex}{(Spatial) Data frame. Land Cover data by hexagon. If
non-spatial, will be converted to spatial sf data frame using the \code{crs} and
\code{coords} arguments. Requires columns identifying the Hex ID as well as the
Stratum ID (see \code{hex_id} and \code{stratum_id} respectively).}

\item{stratum_id}{Column. Identifies larger area (e.g., \code{StudyAreaID} or
\code{Province}).}

\item{hex_id}{Columns. Identifies hexagon ids (e.g., \code{hex_id} or \code{ET_Index}).}

\item{costs}{Data frame. Costs for each hexagon in a RawCost format.}

\item{crs}{Numeric, character, or sf/sfc. Coordinate reference system. Must
be valid input to \code{sf::st_crs()}.}

\item{coords}{Character vector. Names of the columns containing X and Y
coordinates (default \code{c("lon", "lat")}).}

\item{seed}{Numeric. Random seed to use for random sampling. Seed only
applies to specific sampling events (does not change seed in the
environment). \code{NULL} does not set a seed.}
}
\description{
Common arguments and documentation for various functions
}
\details{
Use \verb{@inheritParams common_docs} to include the above in any function
documentation with a matching argument (will only include matching args)
}
\keyword{internal}
