% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_study_area.R
\name{create_study_area}
\alias{create_study_area}
\title{Create Study Area}
\usage{
create_study_area(landscape, study_area_diam, hex_diam, output = "small")
}
\arguments{
\item{landscape}{larger area}

\item{study_area_diam}{study area diameter}

\item{hex_diam}{hexagon diameter}

\item{output}{output type}
}
\value{

}
\description{
Create Study Area
}
