% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_study_area.R
\name{create_study_area}
\alias{create_study_area}
\title{Create Hexagonal grid}
\usage{
create_study_area(
  landscape,
  hexagon_size,
  units,
  HexagonID_label,
  HexagonID_prefix
)
}
\arguments{
\item{landscape}{larger area to create hexagonal grid over.must be sf object}

\item{hexagon_size}{size of hexagon in area or diameter in units (see variable units for options)}

\item{units}{Units of hexagonal units in area or diameter (valid values include 'm2', 'ha', 'km2','km', 'm')}

\item{HexagonID_label}{Name of column for hexigon id}

\item{HexagonID_prefix}{Prefix to affix to hexagon ID}
}
\value{
Returns a sf polygon layer of hexagons with unique IDs
}
\description{
Function that takes a landscape as an sf object and returns a hexagonal grid of a given size.
Allowed inputs are hectres ('ha'), metres squared ('m2'), metres or kilometres ('m', or 'km')
as the diameter of each hexagon
}
