% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASS_grts.R
\name{draw_random_samples}
\alias{draw_random_samples}
\title{Draw random sample}
\usage{
draw_random_samples(
  att_cleaned,
  att.sf,
  num_runs,
  nsamples,
  use_grts = TRUE,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{att_cleaned}{The attribute file for hexagons. Must include Habitat Layers and hexid}

\item{att.sf}{The simple features shape.file object}

\item{num_runs}{The number of times to draw random samples from the attribute files}

\item{nsamples}{The number of samples to draw from each run}

\item{use_grts}{Logical. Indicates if should draw using GRTS or just random sample without
spatial dispersion.}
}
\value{
Returns a sample output in long and wide formats.
}
\description{
Draw random sample
}
