% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASS_grts.R
\name{draw_random_samples}
\alias{draw_random_samples}
\title{Draw random sample}
\usage{
draw_random_samples(
  att_sf,
  num_runs,
  n_samples,
  use_grts = TRUE,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{att_sf}{Spatial Data frame. sf attributes table.}

\item{num_runs}{Numeric. Number of times to draw random samples.}

\item{n_samples}{Numeric. Number of samples to draw from each run.}

\item{use_grts}{Logical. Whether to use GRTS or just sample randomly without
spatial dispersion.}
}
\value{
Samples as spatial data frame.
}
\description{
Draw random sample
}
