% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASS_extract_lcc_cost.R
\name{extract_habitat_cost}
\alias{extract_habitat_cost}
\title{Extract Habitat and Cost}
\usage{
extract_habitat_cost(
  number_iterations,
  n_samples_per_iter,
  sample_hexes,
  study_area_hexes,
  id,
  id_col,
  hab_rast_location,
  primary_roads,
  secondary_roads,
  return_all_,
  write_hexes = F,
  load_hexes = T,
  rds.loc = "output",
  sa.rast.loc = "output",
  quick = T
)
}
\arguments{
\item{number_iterations}{Number of iterations to draw samples from full GRTS}

\item{n_samples_per_iter}{Number of samples to pull per iteration}

\item{sample_hexes}{Sample hexagon file}

\item{study_area_hexes}{Study area hexagon files}

\item{id}{Id of the study area of interest}

\item{id_col}{Column identifying study area}

\item{hab_rast_location}{Location of raster for habitat}

\item{primary_roads}{location of shp file for primary roads}

\item{secondary_roads}{location of shp file for secondary roads}

\item{return_all_}{Logical to return a full results or just the inclusion probabilities}

\item{quick}{Run using cpp}
}
\value{

}
\description{
Extract Habitat and Cost
}
