% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASS_extract_lcc_cost.R
\name{extract_habitat_cost}
\alias{extract_habitat_cost}
\title{Extract Habitat and Cost}
\usage{
extract_habitat_cost(
  number_iterations,
  n_samples_per_iter,
  sample_hexes,
  study_area_hexes,
  id,
  id_col,
  hab_rast_location,
  shape_file_list,
  return_all_,
  nARUs = 30,
  hexid_col = SampleUnitID,
  calc_cost = F,
  calc_hab = F,
  write_hexes = F,
  load_hexes = T,
  rds.loc = "output",
  sa.rast.loc = "output",
  quick = T
)
}
\arguments{
\item{number_iterations}{Number of iterations to draw samples from full GRTS}

\item{n_samples_per_iter}{Number of samples to pull per iteration}

\item{sample_hexes}{Sample hexagon file}

\item{study_area_hexes}{Study area hexagon files}

\item{id}{Id of the study area of interest}

\item{id_col}{Column identifying study area}

\item{hab_rast_location}{Location of raster for habitat}

\item{shape_file_list}{named list of shapefiles used in cost calculation. The current iteration must include the following names:
\describe{
    \item{primary_roads}{sf polygon with primary roads buffer}
    \item{secondary_roads}{sf polygon with atv roads buffer}
    \item{winter_roads}{ sf polygon with winter roads buffer}
    \item{total_roads}{ sf polygon with all roads buffer}
    \item{airport_locations}{ sf points with airport locations}
    \item{camp_locations}{ sf points with camp locations}
    }}

\item{return_all_}{Logical to return a full results or just the inclusion probabilities}

\item{quick}{Run using cpp}
}
\value{

}
\description{
Extract Habitat and Cost
}
