% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASS_full_run.R
\name{full_BASS_run}
\alias{full_BASS_run}
\title{A full BASS run}
\usage{
full_BASS_run(
  num_runs,
  nsamples,
  att,
  att.sp,
  cost,
  return_all = F,
  seed_ = as.integer(Sys.time()),
  HexID_ = HEX100,
  stratumID = StudyAreaID,
  q = F,
  non_ran_set = NULL
)
}
\arguments{
\item{num_runs}{The number of times to draw random samples from hexagons}

\item{nsamples}{The number of samples to draw in each sample}

\item{att}{the attribute table correctly formatted}

\item{att.sp}{the shape file from the attribute table}

\item{cost}{the cost table for each hexagon id}

\item{return_all}{return each piece of BASS implementation}

\item{seed_}{random seed to use for random grts samples}

\item{HexID_}{Column for hexagon id}

\item{stratumID}{Column for larger area id. Likely StudyAreaID or Province}

\item{q}{Run the benefit calculation quickly}

\item{non_ran_set}{Non random set that is added to the hypothetical sample set in benefit calculation.}
}
\value{
a table with inclusion probabilities
}
\description{
A full BASS run
}
