% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASS_studyArea_analysis.R
\name{getresults_BASS}
\alias{getresults_BASS}
\title{Get the results from a BASS grts run}
\usage{
getresults_BASS(grts_output, study_area_results, nARUs)
}
\arguments{
\item{nARUs}{}
}
\value{

}
\description{
Get the results from a BASS grts run
}
