% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASSr-package.R
\docType{data}
\name{lcc2015_codes}
\alias{lcc2015_codes}
\title{2015 National Landcover Classification Table}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
lcc2015_codes
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
