% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASS_full_run.R
\name{noGRTS_BASS_run}
\alias{noGRTS_BASS_run}
\title{A calculate BASS from random samples}
\usage{
noGRTS_BASS_run(
  att_sf,
  samples,
  num_runs,
  n_samples,
  costs,
  seed = as.integer(Sys.time())
)
}
\arguments{
\item{att_sf}{Spatial Data frame. sf attributes table.}

\item{samples}{Hypothetical sample set}

\item{num_runs}{The number of times to draw random samples from hexagons}

\item{n_samples}{The number of samples to draw in each sample}

\item{costs}{the cost table for each hexagon id}

\item{seed}{random seed to use for random grts samples}
}
\value{
a table with inclusion probabilities
}
\description{
A calculate BASS from random samples
}
