% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASS_full_run.R
\name{noGRTS_BASS_run}
\alias{noGRTS_BASS_run}
\title{A calculate BASS from random samples}
\usage{
noGRTS_BASS_run(
  land_hex,
  samples,
  num_runs,
  n_samples,
  costs,
  crs = 4326,
  coords = c("lon", "lat"),
  seed = as.integer(Sys.time()),
  quiet = FALSE
)
}
\arguments{
\item{land_hex}{(Spatial) Data frame. Land Cover data by hexagon. If
non-spatial, will be converted to spatial sf data frame using the \code{crs} and
\code{coords} arguments. Requires columns identifying the Hex ID as well as the
Stratum ID (see \code{hex_id} and \code{stratum_id} respectively).}

\item{samples}{Hypothetical sample set}

\item{num_runs}{The number of times to draw random samples from hexagons}

\item{n_samples}{The number of samples to draw in each sample}

\item{costs}{the cost table for each hexagon id}

\item{crs}{Numeric, character, or sf/sfc. Coordinate reference system. Must
be valid input to \code{sf::st_crs()}.}

\item{coords}{Character vector. Names of the columns containing X and Y
coordinates (default \code{c("lon", "lat")}).}

\item{seed}{random seed to use for random grts samples}
}
\value{
a table with inclusion probabilities
}
\description{
A calculate BASS from random samples
}
