% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ontario}
\alias{ontario}
\title{Polygon SF of Ontario}
\format{
A simple feature data frame with 1 rows and 2 variables:
\describe{
\item{PROV}{Two letter code for Ontario}
\item{NAME}{Name of Province}
\item{geometry}{SF geometry}
}
}
\source{
Map of Canadian Jurisdictions
}
\usage{
ontario
}
\description{
A poloygon with the outline of Ontario
}
\keyword{datasets}
