% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASS_benefit.R
\name{quick_ben}
\alias{quick_ben}
\title{Quick Benefits}
\usage{
quick_ben(
  d,
  samples,
  land_cover_summary,
  hex_id,
  print,
  land_cover_weights = NULL
)
}
\arguments{
\item{d}{Hex data frame - Needs land cover types in columns labelled \code{LC}}

\item{samples}{Hypothetical sample set}

\item{land_cover_summary}{Land cover summary for larger area}

\item{hex_id}{Columns. Identifies hexagon ids (e.g., \code{hex_id} or \code{ET_Index}).}

\item{print}{print details}

\item{land_cover_weights}{a data frame with \code{lc} specifying land cover and
\code{weights} specifying weight.}
}
\value{
data frame with benefits per hex
}
\description{
Quick Benefits
}
