% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASS_benefit.R
\name{quick_ben}
\alias{quick_ben}
\title{Quick Benefits}
\usage{
quick_ben(d, samples, land_cover_summary, col_, pd)
}
\arguments{
\item{d}{Hexagon data.frame}

\item{pd}{print details}
}
\value{

}
\description{
Quick Benefits
}
