% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASS_studyArea_analysis.R
\name{run_grts_on_BASS}
\alias{run_grts_on_BASS}
\title{Run grts sampling on BASSr results}
\usage{
run_grts_on_BASS(
  probs,
  num_runs,
  nARUs,
  os,
  hex_id = NULL,
  stratum_id = NULL,
  remove_hexes = NULL
)
}
\arguments{
\item{probs}{Data frame. Output of \code{calculate_inclusion_probs()} or
\code{full_BASS_run()}.}

\item{num_runs}{Numeric. Number of times to draw random samples.}

\item{nARUs}{Numeric vector. Number of samples or named vector of number of
samples per stratum.}

\item{os}{Numeric vector. Over sample size (proportional) or named vector of
number of samples per stratum.}

\item{hex_id}{Columns. Identifies hexagon ids (e.g., \code{hex_id} or \code{ET_Index})}

\item{stratum_id}{Column. Identifies larger area (e.g., \code{StudyAreaID} or
\code{Province})}

\item{removed_hexes}{Character Vector. Ids of hexagons to remove.}
}
\description{
Run grts sampling on BASSr results
}
