% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASS_studyArea_analysis.R
\name{run_grts_on_BASS}
\alias{run_grts_on_BASS}
\title{Run grts sampling on BASSr results}
\usage{
run_grts_on_BASS(
  n_grts_tests,
  study_area_results,
  nARUs,
  os,
  idcol,
  hexid,
  removedhexes = c("None"),
  return_points = F,
  Stratum = None
)
}
\arguments{
\item{n_grts_tests}{Number of times to draw arus from grts}

\item{study_area_results}{BASSr results}

\item{nARUs}{Number of samples}

\item{os}{Over sample size (proportional)}

\item{idcol}{Identification column for units}

\item{hexid}{Column for hexagon identification}

\item{removedhexes}{Vector of names of hexagons to remove}

\item{return_points}{}
}
\value{

}
\description{
Run grts sampling on BASSr results
}
